pymonitor/requirements.txt
Package                 Version     
----------------------- ------------
amqp                    1.4.9       
anyjson                 0.3.3       
asn1crypto              0.24.0      
bcrypt                  3.1.4       
billiard                3.3.0.23    
celery                  3.1.26.post2
certifi                 2018.4.16   
cffi                    1.11.5      
chardet                 3.0.4       
cryptography            2.2.2       
Django                  2.0.5       
django-filter           1.1.0       
djangorestframework     3.8.2       
djangorestframework-jwt 1.11.0      
idna                    2.6         
kombu                   3.0.37      
oauthlib                2.1.0       
paramiko                2.4.1       
pip                     10.0.1      
psycopg2                2.7.4       
pyasn1                  0.4.2       
pycparser               2.18        
PyJWT                   1.6.3       
PyNaCl                  1.2.1       
PySocks                 1.6.8       
pytz                    2018.4      
redis                   2.10.6      
requests                2.18.4      
requests-oauthlib       0.8.0       
setuptools              39.1.0      
six                     1.11.0      
urllib3                 1.22        
vine                    1.1.4       
wheel                   0.31.0      
